+++ CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed spacing around generics with composite type parameters (e.g. `array of`, `set of`).
+
 ### Changed
 
 - Allow case arm statements to be formatted inline.